# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitor API",
        "description": "API documentation for Uptime Monito project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/monitors/{monitorId}/logs": {
            "get": {
                "tags": [
                    "MonitorLogs"
                ],
                "summary": "Get monitor logs",
                "operationId": "c1bf92f619a30e9feaaad9aa2686cec3",
                "parameters": [
                    {
                        "name": "monitorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "response_time_only",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of monitor logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "MonitorLogs"
                ],
                "summary": "Create a monitor log",
                "operationId": "f022708856d5aef04e78430498e8490c",
                "parameters": [
                    {
                        "name": "monitorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MonitorLog"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Monitor log created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MonitorLog"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logs/{monitorLog}": {
            "get": {
                "tags": [
                    "MonitorLogs"
                ],
                "summary": "Get monitor log detail",
                "operationId": "3e26603a77374a9cf0cb187567f72478",
                "parameters": [
                    {
                        "name": "monitorLog",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor log detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MonitorLog"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "MonitorLogs"
                ],
                "summary": "Update monitor log",
                "operationId": "3402eaa8dc7b6494e5eadf1bfefa425e",
                "parameters": [
                    {
                        "name": "monitorLog",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MonitorLog"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Monitor log updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MonitorLog"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "MonitorLogs"
                ],
                "summary": "Delete monitor log",
                "operationId": "e144d8f080b25737a3d08a69823e22d6",
                "parameters": [
                    {
                        "name": "monitorLog",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Monitor log deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MonitorLog": {
                "title": "MonitorLog",
                "required": [
                    "monitor_id",
                    "started_at",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "monitor_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-11T12:00:00Z"
                    },
                    "status": {
                        "type": "string",
                        "example": "succeeded"
                    },
                    "response_time_ms": {
                        "type": "integer",
                        "example": 123
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}